+++ zeppelin-web/.eslintrc
@@ -14,6 +14,7 @@
   },
   "globals": {
     "angular": false,
+    "inject": false,
     "_": false,
     "jQuery": false,
     "hljs": false,
