+++ zeppelin-web/karma.conf.js
@@ -1,9 +1,26 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 // Karma configuration
 // http://karma-runner.github.io/0.12/config/configuration-file.html
 // Generated on 2014-08-29 using
 // generator-karma 0.8.3
 
-var webpackConfig = require('../webpack.config');
+var webpackConfig = require('./webpack.config');
 
 module.exports = function(config) {
   'use strict';
@@ -13,7 +30,7 @@ module.exports = function(config) {
     autoWatch: true,
 
     // base path, that will be used to resolve files and exclude
-    basePath: '../',
+    basePath: './',
 
     // testing framework to use (jasmine/mocha/qunit/...)
     frameworks: ['jasmine'],
@@ -75,7 +92,7 @@ module.exports = function(config) {
 
       'src/index.js',
       // 'test/spec/**/*.js',
-      {pattern: 'test/spec/**/*.js', watched: false},
+      {pattern: 'src/**/*.test.js', watched: false},
     ],
 
     // list of files / patterns to exclude
@@ -108,12 +125,12 @@ module.exports = function(config) {
     preprocessors: {
       'src/*/{*.js,!(test)/**/*.js}': 'coverage',
       'src/index.js': ['webpack', 'sourcemap',],
-      'test/spec/**/*.js': ['webpack', 'sourcemap',],
+      'src/**/*.test.js': ['webpack', 'sourcemap',],
     },
 
     coverageReporter: {
       type: 'html',
-      dir: '../reports/zeppelin-web-coverage',
+      dir: './reports/zeppelin-web-coverage',
       subdir: '.'
     },
