+++ zeppelin-web/src/components/navbar/navbar.controller.test.js
